+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/items/copper/CopperExpansion.java
@@ -118,6 +118,8 @@ public class CopperExpansion extends Feature {
 			return;
 
 		int y = getNormalizedY(event.getEntity().getBlockY(), level);
+		if (y < 0)
+			return;
 		event.setNewSpeed((float) (event.getNewSpeed() + (0.24f * (Math.pow(y, 0.655f)))));
 	}
 
@@ -135,6 +137,8 @@ public class CopperExpansion extends Feature {
 		int amount = event.getAmount();
 		int newAmount = 0;
 		int y = getNormalizedY(event.getEntity().getBlockY(), level);
+		if (y < 0)
+			return;
 		double chance = 1 - 1 / (1 + 0.37f * Math.pow(y, 0.67f));
 		for (int i = 0; i < amount; i++) {
 			if (event.getRandom().nextFloat() >= chance)
@@ -144,12 +148,11 @@ public class CopperExpansion extends Feature {
 	}
 
 	public static int getNormalizedY(int y, Level level) {
-		if (y > level.getSeaLevel())
-			return y;
 		int startingY = level.getSeaLevel() + deepBonusSeaLevelRelative;
-		//Normalize y to go from 0 at sea level relative to world depth (0~128 usually)
-		y = (startingY - level.getMinBuildHeight()) - (y + startingY);
-		return y;
+		if (y > startingY)
+			return -1;
+		//Normalize y to go from 0 at sea level + deepBonusSeaLevelRelative
+		return (y - startingY) * -1;
 	}
 
 	@SubscribeEvent