+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+### Fixes
+* Copper depth bonus being completely broken
+
 ## 1.22.3
 ### Changes
 * Enduring now adds 50 durability per level to Shields instead of 100